feat: Check for admin user in metods

+++ Bachelor_backend/Controller/AdminController.cs
@@ -41,7 +41,7 @@ namespace Bachelor_backend.Controller
 
             if (success)
             {
-                HttpContext.Session.SetString("UserSession", user.Username);
+                HttpContext.Session.SetString(_loggedIn, user.Username);
                 return Ok(true);
             }
             return Unauthorized("Wrong username or password");
@@ -58,8 +58,14 @@ namespace Bachelor_backend.Controller
         public async Task<ActionResult> RegisterAdmin(AdminUser user)
         {
             var sessionString = HttpContext.Session.GetString(_loggedIn);
-
             if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
+            
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
             {
                 return Unauthorized();
             }
@@ -85,6 +91,19 @@ namespace Bachelor_backend.Controller
         [HttpPost]
         public async Task<ActionResult> CreateTag([FromQuery]string text)
         {
+            var sessionString = HttpContext.Session.GetString(_loggedIn);
+            if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
+            
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
+            {
+                return Unauthorized();
+            }
+            
             bool success = await _textRep.CreateTag(text);
             if (success)
             {
@@ -103,6 +122,19 @@ namespace Bachelor_backend.Controller
         [HttpGet]
         public async Task<ActionResult> GetTags()
         {
+            var sessionString = HttpContext.Session.GetString(_loggedIn);
+            if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
+            
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
+            {
+                return Unauthorized();
+            }
+            
             var list = await _textRep.GetAllTags();
             return Ok(list);
         }
@@ -139,14 +171,26 @@ namespace Bachelor_backend.Controller
         [HttpPost]
         public async Task<ActionResult> CreateText([FromBody] Text text)
         {
+            var sessionString = HttpContext.Session.GetString(_loggedIn);
+            if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
             
-                bool success = await _textRep.CreateText(text);
-                if(success)
-                {
-                    return Ok(true);
-                }
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
+            {
+                return Unauthorized();
+            }
+            
+            bool success = await _textRep.CreateText(text);
+            if(success)
+            {
+                return Ok(true);
+            }
                 
-                return StatusCode(StatusCodes.Status500InternalServerError, false);
+            return StatusCode(StatusCodes.Status500InternalServerError, false);
         }
         
         /// <summary>
@@ -158,6 +202,19 @@ namespace Bachelor_backend.Controller
         [HttpGet]
         public async Task<ActionResult> GetAllTexts()
         {
+            var sessionString = HttpContext.Session.GetString(_loggedIn);
+            if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
+            
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
+            {
+                return Unauthorized();
+            }
+            
             var list = await _textRep.GetAllTexts();
             if(list != null)
             {
@@ -174,6 +231,19 @@ namespace Bachelor_backend.Controller
         [HttpDelete]
         public async Task<ActionResult> DeleteText(int TextId)
         {
+            var sessionString = HttpContext.Session.GetString(_loggedIn);
+            if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
+            
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
+            {
+                return Unauthorized();
+            }
+            
             bool success = await _textRep.DeleteText(TextId);
             if (success)
             {
@@ -191,6 +261,18 @@ namespace Bachelor_backend.Controller
         [HttpDelete]
         public async Task<ActionResult> DeleteTag(int TagId)
         {
+            var sessionString = HttpContext.Session.GetString(_loggedIn);
+            if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
+            
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
+            {
+                return Unauthorized();
+            }
             
             bool success = await _textRep.DeleteTag(TagId);
             if (success)
@@ -208,6 +290,19 @@ namespace Bachelor_backend.Controller
         [HttpGet]
         public async Task<ActionResult> GetNumberOfTexts()
         {
+            var sessionString = HttpContext.Session.GetString(_loggedIn);
+            if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
+            
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
+            {
+                return Unauthorized();
+            }
+            
             int total = await _textRep.GetNumberOfTexts();
             if(total > -1)
             {
@@ -222,16 +317,28 @@ namespace Bachelor_backend.Controller
         /// <returns></returns>
 
         [HttpGet]
-        public async Task<ActionResult> GetNumberOfRecordings(){
+        public async Task<ActionResult> GetNumberOfRecordings()
+        {
+            var sessionString = HttpContext.Session.GetString(_loggedIn);
+            if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
+            
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
+            {
+                return Unauthorized();
+            }
+            
             int total = await _voicerep.GetNumberOfRecordings();
             if (total > -1)
             {
                 return Ok(total);
             }
-            else
-            {
-                return StatusCode(StatusCodes.Status500InternalServerError, -1);
-            }
+            
+            return StatusCode(StatusCodes.Status500InternalServerError, -1);
         }
         /// <summary>
         /// 
@@ -241,6 +348,19 @@ namespace Bachelor_backend.Controller
         [HttpGet]
         public async Task<ActionResult> GetNumberOfUsers()
         {
+            var sessionString = HttpContext.Session.GetString(_loggedIn);
+            if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
+            
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
+            {
+                return Unauthorized();
+            }
+            
             int total = await _textRep.GetNumberOfUsers();
             if(total > -1) {
                 return Ok(total);
@@ -256,6 +376,19 @@ namespace Bachelor_backend.Controller
         [HttpGet]
         public async Task<ActionResult> GetOneText(int id) 
         {
+            var sessionString = HttpContext.Session.GetString(_loggedIn);
+            if (sessionString.IsNullOrEmpty())
+            {
+                return Unauthorized();
+            }
+            
+            var adminFound = await _security.FindAdmin(sessionString);
+
+            if (!adminFound)
+            {
+                return Unauthorized();
+            }
+            
             Text text = await _textRep.GetOneText(id);
             if(text != null)
             {
